Moving konduit-serving-vertx:ai.konduit.serving.vertx.settings to konduit-serving-pipeline:ai.konduit.serving.pipeline.settings
