+++ konduit-serving-pipeline/src/main/java/ai/konduit/serving/pipeline/settings/constants/PropertiesConstants.java
@@ -16,61 +16,54 @@
  *  *****************************************************************************
  */
 
-package ai.konduit.serving.vertx.settings.constants;
+package ai.konduit.serving.pipeline.settings.constants;
 
-import ai.konduit.serving.vertx.config.InferenceConfiguration;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 /**
- * This class contains important constants for different environment variable settings
+ * This class contains important constants for different system properties
  * for konduit-serving.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class EnvironmentConstants {
+public class PropertiesConstants {
 
     /**
-     * Environment variable name for storing port number for the konduit server.
-     * This variable will be prioritizes over {@link InferenceConfiguration#port()}
-     */
-    public static final String KONDUIT_SERVING_PORT = "KONDUIT_SERVING_PORT";
-
-    /**
-     * An environment variable for setting the working directory for konduit serving.
+     * For setting the working directory for konduit serving.
      * The working directory contains the runtime files generated by vertx or
      * konduit-serving itself. The runtime files could contain logs,
      * running process details, vertx cache files etc.
      */
-    public static final String WORKING_DIR = "KONDUIT_WORKING_DIR";
+    public static final String WORKING_DIR = "konduit.working.dir";
 
     /**
-     * Environment variable specifying vertx runtime and cache directory.
+     * System property specifying a directory for keeping vertx runtime and cache files.
      */
-    public static final String VERTX_DIR = "KONDUIT_VERTX_DIR";
+    public static final String VERTX_DIR = "konduit.vertx.dir";
 
     /**
-     * Environment variable specifying build data directory where build logs for the build CLI are kept.
+     * System property specifying build data directory where files for the build CLI are kept.
      */
-    public static final String BUILD_DIR = "KONDUIT_BUILD_DIR";
+    public static final String BUILD_DIR = "konduit.build.dir";
 
     /**
-     * Environment variable specifying profiles data directory where details of individual profiles are kept.
+     * System property specifying profiles data directory where details of individual profiles are kept.
      */
-    public static final String PROFILES_DIR = "KONDUIT_PROFILES_DIR";
+    public static final String PROFILES_DIR = "konduit.profiles.dir";
 
     /**
-     * This variable is responsible for setting the path where the log files for a konduit server
+     * This system property is responsible for setting the path where the log files for a konduit server
      * is kept for the `/logs` endpoint.
      */
-    public static final String ENDPOINT_LOGS_DIR = "KONDUIT_ENDPOINT_LOGS_DIR";
+    public static final String ENDPOINT_LOGS_DIR = "konduit.endpoint.logs.dir";
 
     /**
      * Default directory for containing the command line logs for konduit-serving
      */
-    public static final String COMMAND_LOGS_DIR = "KONDUIT_COMMAND_LOGS_DIR";
+    public static final String COMMAND_LOGS_DIR = "konduit.command.logs.dir";
 
     /**
      * Sets the directory where the file uploads are kept for Vertx BodyHandler
      */
-    public static final String FILE_UPLOADS_DIR = "KONDUIT_FILE_UPLOADS_DIR";
+    public static final String FILE_UPLOADS_DIR = "konduit.file.uploads.dir";
 }
