+++ konduit-serving-vertx-protocols/konduit-serving-http/src/main/java/ai/konduit/serving/vertx/protocols/http/verticle/InferenceVerticleHttp.java
@@ -29,17 +29,15 @@ import ai.konduit.serving.vertx.protocols.http.api.ErrorResponse;
 import ai.konduit.serving.vertx.protocols.http.api.HttpApiErrorCode;
 import ai.konduit.serving.vertx.protocols.http.api.InferenceHttpApi;
 import ai.konduit.serving.vertx.protocols.http.api.KonduitServingHttpException;
-import ai.konduit.serving.vertx.settings.DirectoryFetcher;
-import ai.konduit.serving.vertx.settings.constants.EnvironmentConstants;
+import ai.konduit.serving.pipeline.settings.DirectoryFetcher;
+import ai.konduit.serving.pipeline.settings.constants.EnvironmentConstants;
 import ai.konduit.serving.vertx.verticle.InferenceVerticle;
 import io.vertx.core.Handler;
 import io.vertx.core.Promise;
-import io.vertx.core.http.Http2Settings;
 import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.http.HttpVersion;
 import io.vertx.core.impl.ContextInternal;
 import io.vertx.core.json.JsonObject;
-import io.vertx.core.net.PemKeyCertOptions;
 import io.vertx.ext.web.Route;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
